+++ src/components/NavigationHeader.tsx
@@ -1,0 +1,104 @@
+'use client';
+import { Dialog } from '@headlessui/react';
+import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline';
+import { Link } from 'lucide-react';
+import Image from 'next/image';
+import { useState } from 'react';
+
+import UnderlineLink from '@/components/links/UnderlineLink';
+import UnstyledLink from '@/components/links/UnstyledLink';
+
+import logo from '~/images/ascend_logo.png';
+
+const navigation = [
+  { name: 'Features', href: '#features' },
+  { name: 'Benefits', href: '#benefits' },
+  { name: 'Contact Us', href: '#contact' },
+  { name: 'Company', href: `/company` },
+];
+export const NavigationHeader = () => {
+  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
+  return (
+    <header className='inset-x-0 top-0 z-50'>
+      <nav
+        className='flex items-center justify-between p-6 lg:px-8'
+        aria-label='Global'
+      >
+        <div className='flex lg:flex-1'>
+          <Link href='#' className='-m-1.5 p-1.5'>
+            <span className='sr-only'>Ascend Hire</span>
+            {logo && <Image src={logo} alt='' height={72} width={72} />}
+          </Link>
+        </div>
+        <div className='flex lg:hidden'>
+          <button
+            type='button'
+            className='-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-gray-700'
+            onClick={() => setMobileMenuOpen(true)}
+          >
+            <span className='sr-only'>Open main menu</span>
+            <Bars3Icon className='h-6 w-6' aria-hidden='true' />
+          </button>
+        </div>
+        <div className='hidden lg:flex lg:gap-x-12'>
+          {navigation.map((item) => (
+            <UnderlineLink
+              key={item.name}
+              href={item.href}
+              className='text-sm font-semibold leading-6 text-gray-900'
+            >
+              {item.name}
+            </UnderlineLink>
+          ))}
+        </div>
+        {/* Spacing */}
+        <div className='hidden lg:flex lg:flex-1 lg:justify-end'></div>
+      </nav>
+      <Dialog
+        as='div'
+        className='lg:hidden'
+        open={mobileMenuOpen}
+        onClose={setMobileMenuOpen}
+      >
+        <div className='fixed inset-0 z-50' />
+        <Dialog.Panel className='fixed inset-y-0 right-0 z-50 w-full overflow-y-auto bg-white px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10'>
+          <div className='flex items-center justify-between'>
+            <a href='#' className='-m-1.5 p-1.5'>
+              <span className='sr-only'>Ascend Hire</span>
+              <Image
+                className='h-8 w-auto'
+                src={logo}
+                alt=''
+                height={84}
+                width={84}
+              />
+            </a>
+            <button
+              type='button'
+              className='-m-2.5 rounded-md p-2.5 text-gray-700'
+              onClick={() => setMobileMenuOpen(false)}
+            >
+              <span className='sr-only'>Close menu</span>
+              <XMarkIcon className='h-6 w-6' aria-hidden='true' />
+            </button>
+          </div>
+          <div className='mt-6 flow-root'>
+            <div className='-my-6 divide-y divide-gray-500/10'>
+              <div className='space-y-2 py-6'>
+                {navigation.map((item) => (
+                  <UnstyledLink
+                    key={item.name}
+                    href={item.href}
+                    className='-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50'
+                  >
+                    {item.name}
+                  </UnstyledLink>
+                ))}
+              </div>
+            </div>
+          </div>
+        </Dialog.Panel>
+      </Dialog>
+    </header>
+  );
+};
